+++ src/private/IRremoteBoardDefs.h
@@ -20,14 +20,12 @@
  * board-dependent versions of these macros SENDPIN_ON(pin) and SENDPIN_OFF(pin).
  */
 #ifndef SENDPIN_ON
-/** Board dependent macro to turn on the pin given as argument. */
+/** Default macro to turn on the pin given as argument. */
 #define SENDPIN_ON(pin)  digitalWrite(pin, HIGH)
 #endif
 
 #ifndef SENDPIN_OFF
-/**
- * Board dependent macro to turn off the pin given as argument.
- */
+/** Default macro to turn off the pin given as argument. */
 #define SENDPIN_OFF(pin) digitalWrite(pin, LOW)
 #endif
 
@@ -455,7 +453,7 @@ static void timerConfigForReceive() {
 
 #define TIMER_RESET_INTR_PENDING
 #  if defined(ARDUINO_AVR_PROMICRO) // Sparkfun Pro Micro
-#define TIMER_ENABLE_SEND_PWM       TCNT4 = 0; (TCCR4A |= _BV(COM4A0))     // Use complimentary O̅C̅4̅A̅ output on pin 5
+#define TIMER_ENABLE_SEND_PWM       TCNT4 = 0; (TCCR4A |= _BV(COM4A0))     // Use complementary OC4A output on pin 5
 #define TIMER_DISABLE_SEND_PWM      (TCCR4A &= ~(_BV(COM4A0)))  // (Pro Micro does not map PC7 (32/ICP3/CLK0/OC4A)
                                                             // of ATmega32U4 )
 #  else
@@ -962,68 +960,77 @@ static void timerConfigForReceive() {
  * If defined, denotes pin number of LED that should be blinked during IR reception.
  * Leave undefined to disable blinking.
  */
-#define BLINKLED        LED_BUILTIN
+#define FEEDBACK_LED        LED_BUILTIN
 
 /**
- * Board dependent macro to turn BLINKLED on.
+ * Board dependent macro to turn FEEDBACK_LED on.
  */
-#define BLINKLED_ON()   digitalWrite(BLINKLED, HIGH)
+#define FEEDBACK_LED_ON()   digitalWrite(FEEDBACK_LED, HIGH)
 
 /**
- * Board dependent macro to turn BLINKLED off.
+ * Board dependent macro to turn FEEDBACK_LED off.
  */
-#define BLINKLED_OFF()  digitalWrite(BLINKLED, LOW)
+#define FEEDBACK_LED_OFF()  digitalWrite(FEEDBACK_LED, LOW)
 
 // Sparkfun Pro Micro is __AVR_ATmega32U4__ but has different external circuit
 #elif defined(ARDUINO_AVR_PROMICRO)
 // We have no built in LED -> reuse RX LED
-#define BLINKLED        LED_BUILTIN_RX
-#define BLINKLED_ON()   RXLED1
-#define BLINKLED_OFF()  RXLED0
+#define FEEDBACK_LED        LED_BUILTIN_RX
+#define FEEDBACK_LED_ON()   RXLED1
+#define FEEDBACK_LED_OFF()  RXLED0
 
 // Arduino Leonardo + others
 #elif defined(__AVR_ATmega32U4__)
-#define BLINKLED        LED_BUILTIN
-#define BLINKLED_ON()   (PORTC |= B10000000)
-#define BLINKLED_OFF()  (PORTC &= B01111111)
+#define FEEDBACK_LED        LED_BUILTIN
+#define FEEDBACK_LED_ON()   (PORTC |= B10000000)
+#define FEEDBACK_LED_OFF()  (PORTC &= B01111111)
 
 #elif defined(__AVR_ATmega32U4__) || defined(__AVR_ATmega8U2__) || defined(__AVR_ATmega16U2__)  || defined(__AVR_ATmega32U2__)
-#define BLINKLED        LED_BUILTIN
-#define BLINKLED_ON()   (digitalWrite(LED_BUILTIN, HIGH))
-#define BLINKLED_OFF()  (digitalWrite(LED_BUILTIN, LOW))
+#define FEEDBACK_LED        LED_BUILTIN
+#define FEEDBACK_LED_ON()   (digitalWrite(LED_BUILTIN, HIGH))
+#define FEEDBACK_LED_OFF()  (digitalWrite(LED_BUILTIN, LOW))
 
 // Arduino Uno, Nano etc
 #elif defined(__AVR_ATmega328P__) || defined(__AVR_ATmega328PB__) || defined(__AVR_ATmega168__)
-#define BLINKLED        LED_BUILTIN
-#define BLINKLED_ON()   (PORTB |= B00100000)
-#define BLINKLED_OFF()  (PORTB &= B11011111)
+#define FEEDBACK_LED        LED_BUILTIN
+#define FEEDBACK_LED_ON()   (PORTB |= B00100000)
+#define FEEDBACK_LED_OFF()  (PORTB &= B11011111)
 
 #elif defined(__AVR_ATmega1280__) || defined(__AVR_ATmega2560__)
-#define BLINKLED        13
-#define BLINKLED_ON()   (PORTB |= B10000000)
-#define BLINKLED_OFF()  (PORTB &= B01111111)
+#define FEEDBACK_LED        13
+#define FEEDBACK_LED_ON()   (PORTB |= B10000000)
+#define FEEDBACK_LED_OFF()  (PORTB &= B01111111)
 
 #elif defined(__AVR_ATmega644P__) || defined(__AVR_ATmega644__)
-#define BLINKLED        0
-#define BLINKLED_ON()   (PORTD |= B00000001)
-#define BLINKLED_OFF()  (PORTD &= B11111110)
+#define FEEDBACK_LED        0
+#define FEEDBACK_LED_ON()   (PORTD |= B00000001)
+#define FEEDBACK_LED_OFF()  (PORTD &= B11111110)
 
 // TinyCore boards
 #elif defined(__AVR_ATtiny1616__)  || defined(__AVR_ATtiny3216__) || defined(__AVR_ATtiny3217__)
 // No LED available on the board, take LED_BUILTIN which is also the DAC output
-#define BLINKLED        LED_BUILTIN // PA6
-#define BLINKLED_ON()   (PORTC.OUTSET = _BV(6))
-#define BLINKLED_OFF()  (PORTC.OUTCLR = _BV(6))
+#define FEEDBACK_LED        LED_BUILTIN // PA6
+#define FEEDBACK_LED_ON()   (PORTC.OUTSET = _BV(6))
+#define FEEDBACK_LED_OFF()  (PORTC.OUTCLR = _BV(6))
 
 #elif defined(ESP32)
-// No system LED on ESP32, disable blinking by NOT defining BLINKLED
+// No system LED on ESP32, disable blinking by NOT defining FEEDBACK_LED
 
 #elif defined(PARTICLE)
 
-#define BLINKLED       D7
-#define BLINKLED_ON()  digitalWrite(BLINKLED,1)
-#define BLINKLED_OFF() digitalWrite(BLINKLED,0)
+#define FEEDBACK_LED       D7
+#define FEEDBACK_LED_ON()  digitalWrite(FEEDBACK_LED,1)
+#define FEEDBACK_LED_OFF() digitalWrite(FEEDBACK_LED,0)
 
+// Arduino Zero and BluePill have an LED which is active low
+#elif defined(__STM32F1__) || defined(STM32F1xx)
+#  if defined(LED_BUILTIN)
+#     if !defined(FEEDBACK_LED)
+#define FEEDBACK_LED        LED_BUILTIN
+#    endif
+#define FEEDBACK_LED_ON()   digitalWrite(FEEDBACK_LED, LOW)
+#define FEEDBACK_LED_OFF()  digitalWrite(FEEDBACK_LED, HIGH)
+#  endif
 /*
  * These are the boards for which the default case was verified and the warning below is suppressed
  */
@@ -1032,8 +1039,8 @@ static void timerConfigForReceive() {
 #elif !(defined(__AVR_ATmega4809__) || defined(NRF5) || defined(ARDUINO_ARCH_NRF52840) || defined(__AVR_ATtiny84__) || defined(__AVR_ATtiny85__) \
     || defined(ARDUINO_ARCH_SAMD) \
     || defined(CORE_LED0_PIN) \
-    || defined(__STM32F1__) || defined(ARDUINO_ARCH_STM32F1) \
-    || defined(STM32F1xx) || defined(ARDUINO_ARCH_STM32) \
+    || defined(ARDUINO_ARCH_STM32F1) \
+    || defined(ARDUINO_ARCH_STM32) \
     )
 /*
  * print a warning
@@ -1045,9 +1052,11 @@ static void timerConfigForReceive() {
  * Default case
  */
 #  ifdef LED_BUILTIN
-#define BLINKLED        LED_BUILTIN
-#define BLINKLED_ON()   digitalWrite(BLINKLED, HIGH)
-#define BLINKLED_OFF()  digitalWrite(BLINKLED, LOW)
+#     if !defined(FEEDBACK_LED)
+#define FEEDBACK_LED        LED_BUILTIN
+#    endif
+#define FEEDBACK_LED_ON()   digitalWrite(FEEDBACK_LED, HIGH)
+#define FEEDBACK_LED_OFF()  digitalWrite(FEEDBACK_LED, LOW)
 #  endif
 #endif